pyproject.toml
[project]
name = "covjson-pydantic"
description = "The Pydantic models for CoverageJSON"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "KNMI Data Platform Team", email = "opendata@knmi.nl"},
]
keywords = ["covjson", "Pydantic"]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
    "Typing :: Typed",
]
version = "0.1.0"
dependencies = ["pydantic", "orjson"]

[project.optional-dependencies]
test = ["pytest", "PyHamcrest"]
dev = ["pre-commit"]

[project.urls]
Source = "https://github.com/PaulVanSchayck/pypi-github-test"

[build-system]
requires = ["flit>=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "covjson_pydantic"

[tool.flit.sdist]
exclude = [
    "test/",
    ".github/",
]


[tool.mypy]
plugins = [
  "pydantic.mypy"
]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
show_error_codes = true

[tool.pydantic-mypy]
warn_untyped_fields = true

[tool.pydocstyle]
select = "D1"
match = "(?!test).*.py"
